h/requirements.txt
streamlit
faiss-cpu
numpy
pandas
mistralai
sentence-transformers
python-dotenv